Move into segment subfolder
