+++ visualizer/src/Project/service/edit/segment/index.js
@@ -1,0 +1,3 @@
+import createSegmentMachine from './segmentMachine';
+
+export default createSegmentMachine;
